pyproject.toml
[project]
name = "panel-monitoring-agent"
version = "1.0.0"
description = "Panel Monitoring Agent"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
  "langgraph",
  "langgraph-prebuilt",
  "langgraph-sdk",
  "langgraph-checkpoint-sqlite",
  "langsmith",
  "langchain-community",
  "langchain-core",
  "langchain-openai",
  "langchain-google-vertexai",
  "google-cloud-aiplatform",
  "google-api-core",
  "python-dotenv",
  "pydantic",
  "pydantic-settings",
  "notebook",
  "tavily-python",
  "wikipedia",
  "trustcall",
  "langgraph-cli[inmem]",
  "langchain-google-genai",
  "google-auth>=2.40.3",
  "ruff>=0.13.2",
  "pytest>=8.4.2",
]


[project.optional-dependencies]
dev = [
  "ruff>=0.13.2",
  "pytest>=8.4.2",
  "langchain>=0.3.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
managed = true
package = false

[tool.ruff]
exclude = ["**/*.ipynb"]
